Downloads/Tsunade/Tsunade/backend_backup/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# CORS middleware
fastapi-cors==0.0.6

# OCR dependencies
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.8.1.78

# Image processing
numpy==1.24.3

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# File handling
python-multipart==0.0.6

# OpenAI API
openai==0.28.1